+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# ...
+
+* Adapter `PILFile` now accepts np.ndarray image
+* `TableStoreFiledir` now supports `insert_rows` and `delete_rows` in case multiply suffixes in template
+
 # 0.13.8
 
 * Fix [#301](https://github.com/epoch8/datapipe/issues/301)